Check in dist files
This is a fork of https://github.com/react-dnd/react-dnd
Starting in an unmodified state, I built it with `yarn install`
I then copied packages/*/dist/*.js into dist/.
In order to make this work I had to make minor changes
to .gitignore and to package.json (so it wouldn't run prettier
on the minified dist files). I'm also checking in the modified
yarn.lock file.